#203599 #207247
Interpolation library unit tests have been created.
Some warnings have been fixed.

(Please do not delete my branch when completing/merging it)

Related work items: #203599, #207247

+++ development/libraries/Interpolation/src/RBFInterpolant.h
@@ -135,8 +135,8 @@ Numerics::RBFInterpolant<RadialBasisFunction>::RBFInterpolant
      const int                 initialPolynomialDegree ) :
 
      interpolationPoints ( newInterpolationPoints ),
-     coefficients ( newInterpolationPoints.size () + numberOfPolynomialTerms<Point::DIMENSION>( initialPolynomialDegree )),
-     workSpace    ( newInterpolationPoints.size () + numberOfPolynomialTerms<Point::DIMENSION>( initialPolynomialDegree )) {
+     coefficients ( (int)newInterpolationPoints.size () + numberOfPolynomialTerms<Point::DIMENSION>( initialPolynomialDegree )),
+     workSpace    ( (int)newInterpolationPoints.size () + numberOfPolynomialTerms<Point::DIMENSION>( initialPolynomialDegree )) {
 
 
   polynomialDegree = initialPolynomialDegree;
@@ -176,7 +176,7 @@ const Numerics::Vector& Numerics::RBFInterpolant<RadialBasisFunction>::getCoeffi
 template <class RadialBasisFunction>
 void Numerics::RBFInterpolant<RadialBasisFunction>::setInterpolationPoints ( const PointArray& newPoints ) {
   interpolationPoints = newPoints;
-  workSpace.resize ( interpolationPoints.size () + numberOfPolynomialTerms<Point::DIMENSION>( polynomialDegree ));
+  workSpace.resize ( static_cast<int>(interpolationPoints.size()) + numberOfPolynomialTerms<Point::DIMENSION>( polynomialDegree ));
 }
 
 //------------------------------------------------------------//
